Added more inputs into details page Setup the Geoloc and Weather calls Did some Data cleanup Added some classes for future experiment

+++ WhatToDo/WhatToDo/ViewModel/MainViewModel.cs
@@ -1,10 +1,5 @@
 ﻿
 
-using System.Linq;
-using WhatToDo.Service;
-using WhatToDo.Data;
-using WhatToDo.View;
-
 namespace WhatToDo.ViewModel;
 
 public partial class MainViewModel : BaseViewModel
@@ -12,8 +7,9 @@ public partial class MainViewModel : BaseViewModel
     IDataService dataService;
     IDialogService dialogService;
     IGeolocation geolocation;
-
-    public Session Session { get; set; }
+    IConnectivity connectivity;
+    ISessionService session;
+    WeatherService weatherService;
 
     public ObservableCollection<ToDoItem> Items { get; } = new ();
 
@@ -23,15 +19,38 @@ public partial class MainViewModel : BaseViewModel
     [ObservableProperty]
     ToDoItem selectedItem;
 
-
-
-    public MainViewModel(IDataService dataService, IDialogService dialogService, IGeolocation geolocation)
+    public MainViewModel(
+        IDataService dataService, IDialogService dialogService, 
+        IGeolocation geolocation, ISessionService session, IConnectivity connectivity, WeatherService weatherService)
     {
         Title = "WhatToDo";
         SelectedItem = new ToDoItem();
         this.dataService = dataService;
         this.dialogService = dialogService;
         this.geolocation = geolocation;
+        this.connectivity = connectivity;
+        this.weatherService = weatherService;
+        this.session = session;
+        Task.FromResult(InitialLoad());
+    }
+
+    public async Task InitialLoad()
+    {
+        if (session.CurrentGeolocInfo.Location == null || DateTime.Now >= session.CurrentGeolocInfo.NextUpdateTime)
+        {
+            await GetGeoLocAsync();
+        }
+
+        if (session.CurrentGeolocInfo.Location != null && (session.WeatherForcast == null || session.WeatherForcast.Count == 0))
+        {
+            await GetWeatherAsync();
+            await dataService.SaveWeatherData();
+        }
+
+        if (Items == null || Items.Count == 0)
+        {
+            await GetItemsAsync();
+        }
     }
 
     #region Navigation
@@ -46,10 +65,11 @@ public partial class MainViewModel : BaseViewModel
     [RelayCommand]
     async Task RefreshAsync()
     {
+
         //Get Todo Item Data
         if (Items == null || Items.Count == 0)
         {
-            await GetDataAsync();
+            await GetItemsAsync();
         }
         else
         {
@@ -60,20 +80,15 @@ public partial class MainViewModel : BaseViewModel
             SelectedItem = new ToDoItem();
         }
 
-        //Get Session Data
-        if (Session == null)
-        {
-            //TODO: Standardize Session and Data Service Call
-            Session = await dataService.GetSession();
-        }
         DisplayItems = null;
         DisplayItems = Items;
         IsRefreshing = false;
     }
     #endregion UI
 
+    #region Get Data
     [RelayCommand]
-    async Task GetDataAsync()
+    async Task GetItemsAsync()
     {
         if (IsBusy) return;
         try
@@ -102,14 +117,14 @@ public partial class MainViewModel : BaseViewModel
         }
     }
 
-    public override async Task Cache()
+/*    public override async Task Cache()
     {
-        await dataService.SaveSession(Session);
+        await dataService.SaveSession(session);
         return;
-    }
+    }*/
 
     [RelayCommand]
-    async Task GetGeoLoc()
+    async Task GetGeoLocAsync()
     {
         if (IsBusy)
             return;
@@ -118,24 +133,80 @@ public partial class MainViewModel : BaseViewModel
             IsBusy = true;
             // Get cached location, else get real location.
             Location location = await geolocation.GetLastKnownLocationAsync();
-            if (location == null)
+            if (location is null || DateTime.Now > session.CurrentGeolocInfo.NextUpdateTime)
             {
                 location = await geolocation.GetLocationAsync(new GeolocationRequest
                 {
                     DesiredAccuracy = GeolocationAccuracy.Medium,
                     Timeout = TimeSpan.FromSeconds(30)
                 });
+
+                if (location is null)
+                {
+                    throw new ArgumentNullException(nameof(location));
+                }
             }
 
+            session.CurrentGeolocInfo.Location = location;
+
         }
-        catch { }
+        catch (Exception ex)
+        {
+            Debug.WriteLine($"Unable to get Updated Location: {ex.Message}");
+            await dialogService.DisplayAlert("Error!", ex.Message, "OK");
+        }
+        finally { IsBusy = false; } 
     }
+
+    [RelayCommand]
+    async Task GetWeatherAsync()
+    {
+        if (IsBusy)
+            return;
+        try
+        {
+            if (session.WeatherForcast is null || session.WeatherForcast.Count < 14 || session.WeatherForcast.Keys.Any(x => x < DateTime.Now.AddDays(-1)))
+            {
+                if (connectivity.NetworkAccess != NetworkAccess.Internet)
+                {
+                    await dialogService.DisplayAlert("No connectivity!", "Please check internet and try again.", "OK");
+                    return;
+                }
+
+                List<WeatherData> data = await weatherService.GetWeather(session.CurrentGeolocInfo.Location);
+
+                if (data is null || data.Count == 0) 
+                { 
+                    throw new ArgumentNullException(nameof(data)); 
+                }
+                else
+                {
+                    session.WeatherForcast.Clear();
+                    foreach (WeatherData halfDay in data)
+                    {
+                        session.WeatherForcast[halfDay.startTime] = halfDay;
+                    }
+                }
+            }
+        }
+        catch (Exception ex)
+        {
+            Debug.WriteLine($"Unable to get Weather Data: {ex.Message}");
+            await dialogService.DisplayAlert("Error!", ex.Message, "OK");
+        }
+        finally
+        {
+            IsBusy = false; 
+        }
+    }
+    #endregion Get Data
 
     #region ToDo Item Details / Edits
 
     [RelayCommand]
     async Task Tap(ToDoItem item)
     {
+        if (IsBusy) return;
         await Task.FromResult(SelectedItem = item);
         await GoToDetails(SelectedItem);
     }
@@ -143,13 +214,33 @@ public partial class MainViewModel : BaseViewModel
     [RelayCommand]
     async Task AddItem()
     {
+        if (IsBusy) return;
         Items.Add(SelectedItem);
         await GoToDetails(SelectedItem);
     }
 
     [RelayCommand]
+    async Task CompleteAsync(ToDoItem item)
+    {
+        if (IsBusy) return;
+        for (int i = 0; i < Items.Count; i++)
+        {
+            if (Items[i].Id == item.Id)
+            {
+                Items[i].IsComplete = !item.IsComplete;
+                break;
+            }
+        }
+        await dataService.UpdateItemAsync(item);
+        await Task.Delay(200);
+        DisplayItems = null;
+        DisplayItems = Items;        
+    }
+
+    [RelayCommand]
     async Task DeleteAsync(ToDoItem item)
     {
+        if (IsBusy) return;
         if (Items.Contains(item))
         {
             await dataService.DeleteItemAsync(item);
